requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.2
    # via langchain-community
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anthropic==0.73.0
    # via langchain-anthropic
anyio==4.11.0
    # via
    #   anthropic
    #   httpx
    #   openai
async-timeout==4.0.3
    # via
    #   aiohttp
    #   langchain-classic
attrs==25.4.0
    # via aiohttp
beautifulsoup4==4.14.2
    # via wikipedia
brotli==1.2.0
    # via httpx
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   ddgs
    #   duckduckgo-search
colorama==0.4.6
    # via
    #   click
    #   tqdm
dataclasses-json==0.6.7
    # via langchain-community
ddgs==9.10.0
    # via -r requirements.in
distro==1.9.0
    # via
    #   anthropic
    #   openai
docstring-parser==0.17.0
    # via anthropic
duckduckgo-search==8.1.1
    # via -r requirements.in
exceptiongroup==1.3.0
    # via anyio
fake-useragent==2.2.0
    # via ddgs
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.2.4
    # via sqlalchemy
h11==0.16.0
    # via httpcore
h2==4.3.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httpx[brotli,http2,socks]==0.28.1
    # via
    #   anthropic
    #   ddgs
    #   langgraph-sdk
    #   langsmith
    #   openai
httpx-sse==0.4.3
    # via langchain-community
hyperframe==6.1.0
    # via h2
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jiter==0.12.0
    # via
    #   anthropic
    #   openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==1.0.7
    # via -r requirements.in
langchain-anthropic==1.0.4
    # via -r requirements.in
langchain-classic==1.0.0
    # via langchain-community
langchain-community==0.4.1
    # via -r requirements.in
langchain-core==1.0.5
    # via
    #   langchain
    #   langchain-anthropic
    #   langchain-classic
    #   langchain-community
    #   langchain-openai
    #   langchain-text-splitters
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langchain-openai==1.0.3
    # via -r requirements.in
langchain-text-splitters==1.0.0
    # via langchain-classic
langgraph==1.0.3
    # via langchain
langgraph-checkpoint==3.0.1
    # via
    #   langgraph
    #   langgraph-prebuilt
langgraph-prebuilt==1.0.4
    # via langgraph
langgraph-sdk==0.2.9
    # via langgraph
langsmith==0.4.43
    # via
    #   langchain-classic
    #   langchain-community
    #   langchain-core
lxml==6.0.2
    # via
    #   ddgs
    #   duckduckgo-search
marshmallow==3.26.1
    # via dataclasses-json
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.1.0
    # via typing-inspect
numpy==2.2.6
    # via langchain-community
openai==2.8.0
    # via langchain-openai
orjson==3.11.4
    # via
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.12.0
    # via langgraph-checkpoint
packaging==25.0
    # via
    #   langchain-core
    #   langsmith
    #   marshmallow
primp==0.15.0
    # via
    #   ddgs
    #   duckduckgo-search
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
pydantic==2.12.4
    # via
    #   -r requirements.in
    #   anthropic
    #   langchain
    #   langchain-anthropic
    #   langchain-classic
    #   langchain-core
    #   langgraph
    #   langsmith
    #   openai
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.12.0
    # via langchain-community
python-dotenv==1.2.1
    # via
    #   -r requirements.in
    #   pydantic-settings
pyyaml==6.0.3
    # via
    #   langchain-classic
    #   langchain-community
    #   langchain-core
regex==2025.11.3
    # via tiktoken
requests==2.32.5
    # via
    #   langchain-classic
    #   langchain-community
    #   langsmith
    #   requests-toolbelt
    #   tiktoken
    #   wikipedia
requests-toolbelt==1.0.0
    # via langsmith
sniffio==1.3.1
    # via
    #   anthropic
    #   anyio
    #   openai
socksio==1.0.0
    # via httpx
soupsieve==2.8
    # via beautifulsoup4
sqlalchemy==2.0.44
    # via
    #   langchain-classic
    #   langchain-community
tenacity==9.1.2
    # via
    #   langchain-community
    #   langchain-core
tiktoken==0.12.0
    # via langchain-openai
tqdm==4.67.1
    # via openai
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   anthropic
    #   anyio
    #   beautifulsoup4
    #   exceptiongroup
    #   langchain-core
    #   multidict
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspect
    #   typing-inspection
typing-inspect==0.9.0
    # via dataclasses-json
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.5.0
    # via requests
wikipedia==1.4.0
    # via -r requirements.in
xxhash==3.6.0
    # via langgraph
yarl==1.22.0
    # via aiohttp
zstandard==0.25.0
    # via langsmith
